Removed nesting of test builder classes
