+++ DiaryTest/DiaryTest.csproj
@@ -50,11 +50,13 @@
     <Reference Include="System.Data" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="AppointmentBuilder.cs" />
     <Compile Include="AppointmentTest.cs" />
     <Compile Include="DateTest.cs" />
     <Compile Include="DateTimeTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ContactTest.cs" />
+    <Compile Include="ReminderBuilder.cs" />
     <Compile Include="ReminderTest.cs" />
     <Compile Include="CalendarEventTest.cs" />
   </ItemGroup>
